projects/example_project/pip-requirements.txt
#./dist/ml_utils-0.0.1-py3-none-any.whl
snowflake-snowpark-python
snowflake-ml-python
scikit-learn
